Grouping tests and enhance tests setup
