+++ frontend/packages/kubevirt-plugin/integration-tests/tests/vm.action.base.scenario.ts
