+++ frontend/packages/kubevirt-plugin/integration-tests/tests/vm.action.clone.scenario.ts
