+++ frontend/packages/kubevirt-plugin/integration-tests/tests/vm.detail.node-selector.scenario.ts
@@ -1,26 +1,12 @@
 import { browser, ExpectedConditions as until } from 'protractor';
-import { testName } from '@console/internal-integration-tests/protractor.conf';
-import { createResource, deleteResource, click } from '@console/shared/src/test-utils/utils';
+import { click } from '@console/shared/src/test-utils/utils';
 import * as editNodeSelectorView from '../views/editNodeSelectorView';
 import * as virtualMachineView from '../views/virtualMachine.view';
 import { saveButton } from '../views/kubevirtUIResource.view';
 import { VM_CREATE_AND_EDIT_TIMEOUT_SECS } from './utils/consts';
-import { VirtualMachine } from './models/virtualMachine';
-import { getVMManifest } from './utils/mocks';
-import { getRandStr } from './utils/utils';
+import { vm } from './vm.setup.scenario';
 
 describe('KubeVirt VM detail - edit Node Selector', () => {
-  const testVM = getVMManifest('Container', testName, `node-selector-vm-${getRandStr(5)}`);
-  const vm = new VirtualMachine(testVM.metadata);
-
-  beforeAll(async () => {
-    createResource(testVM);
-  });
-
-  afterAll(() => {
-    deleteResource(testVM);
-  });
-
   it(
     'ID(CNV-4133) Adds a Node Selector, then removes it',
     async () => {
