+++ frontend/packages/kubevirt-plugin/integration-tests/tests/vm.tab.consoles.rdp.scenario.ts
