+++ frontend/packages/kubevirt-plugin/integration-tests/tests/vm.tab.consoles.scenario.ts
