+++ frontend/packages/kubevirt-plugin/integration-tests/tests/vm.tab.dashboard.scenario.ts
