+++ frontend/packages/kubevirt-plugin/integration-tests/tests/vm.tab.environment.scenario.ts
