+++ frontend/packages/kubevirt-plugin/integration-tests/tests/vm.teardown.scenario.ts
@@ -1,0 +1,8 @@
+import { VM_ACTION } from './utils/consts';
+import { vm } from './vm.setup.scenario';
+
+describe('KubeVirt VM teardown', () => {
+  it('Kubevirt VM delete', async () => {
+    await vm.action(VM_ACTION.Delete, false);
+  });
+});
